--- srcs/.env
@@ -1,30 +1,0 @@
-NODE_VERSION=latest
-DEV_ENV=dev
-
-PORT=3000
-
-#Services
-WEBSITE_SERVICE=fastify
-DB_SERVICE=sqlite
-MONITORING_SERVICE=prometheus
-MONITORING_VISUALISER_SERVICE=grafana
-
-WEBSITE_SERVICE_IMAGE=fastify
-MONITORING_SERVICE_IMAGE=prom/prometheus
-MONITORING_VISUALISER_SERVICE_IMAGE=grafana/grafana-oss
-
-# Locations
-SRCS=./services/
-
-VOLUME_DB=/goinfre/$USER/pong/data/sqlite/
-VOLUME_SCRIPTS=/goinfre/$USER/pong/data/fastify/
-
-# Volume in containers
-LOCATION_DB=/home/db/
-LOCATION_SCRIPTS=/home/fastify/
-
-# Secrets
-SECRETS=./secrets/
-
-SSL_CERTIFICATE=ssl.crt
-SSL_KEY=ssl.key
