--- srcs/secrets/ssl.crt
@@ -1,32 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFfzCCA2egAwIBAgIUdDHJ6Z/CFXjP09R40M9M1OFDKH0wDQYJKoZIhvcNAQEL
-BQAwTzELMAkGA1UEBhMCRlIxDzANBgNVBAgMBkZyYW5jZTERMA8GA1UEBwwITXVs
-aG91c2UxDTALBgNVBAoMBHBvbmcxDTALBgNVBAMMBG5vbmUwHhcNMjUwNjE4MDgz
-MTU5WhcNMjUwNzE4MDgzMTU5WjBPMQswCQYDVQQGEwJGUjEPMA0GA1UECAwGRnJh
-bmNlMREwDwYDVQQHDAhNdWxob3VzZTENMAsGA1UECgwEcG9uZzENMAsGA1UEAwwE
-bm9uZTCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBALoOcGdBwu+wdmLh
-i8SNAtB1A6EvuwC0Hex+HBqVyupIdvhCUKe4Lf9v9ygp2iByR3j07lDKzIUfANsg
-X9zNvInX9ESbGn+KYct1Kk/nY27BRipUB1IqXXqTb/0xZ/tH8npFptrupX0/br9W
-dw9n9zbIUQ5AyKH+/llwpnEnJyRkmqQseKprhGmNA7G+0oD1C4ItkwGpJ1msWsDf
-92P4ASUxgO4GyBnINprY+jUCJAg1rioG2CSEBVXVoWAyXoITD1RFUac/W6yn3jZq
-+XzDC12KLLpJ1mldpc+pAjC89xX5XgJzuQeqru73Py6v/LF54wZKcc/0M3A74ytk
-IKOPc3x7tOvwExVc47N1PbxCO9UCojVSA77Mj9qwsvDWMLHKUFwkoNVbPFb8psB2
-Hn8HzLtj4WrXj1FOl0vHOfXMVXDmVmYaKvDnfTumRGQGGS8LpGiDu99ByrbsffYB
-gjbsaPsAgvZUJ9U7Njk4oiba5gv/X11jQ/dV91Px+8sF9ks/9oM7Ghwgv67lB8lj
-m4GIoapKCqBW6XZeHQL8VOHkeCjZwtTZ2buxRLko3FoHIFtkuW8o1pMNMRYyGlsR
-lq1hIDYkJ6jR69oYCzcpyP1/ME2W3Vi+ev1IVwn5ATnxAiQ77Yc2/VnaSx6ExFOt
-Gz7owiUzsaJPU94IJx77Elpi2ijnAgMBAAGjUzBRMB0GA1UdDgQWBBRksrqgGkT7
-5jY1QTVTO984D4EIrTAfBgNVHSMEGDAWgBRksrqgGkT75jY1QTVTO984D4EIrTAP
-BgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4ICAQA/d0T13Q+WGxp2znJU
-k2qokjDASJNuGlSeYjNDqjzulayYQv6lxxK1tRXyNGxCdWr6Qpdzau0xP2i//g1M
-a2mOyuY9sH7gATBSi4Q8OMvqulfDYt6+DQ4jLu9T0cmZm0MHJI3DMJQt76rWiL6u
-zEKit35j9bNa9VBaA5q+24sOV/IW0fqhxVFQLa4IMO7IkY+cdoT4Kp9Sqdhse8ah
-JenH9WZxoMmtl88dU0qhqRe5QnK2LSGJJuRF+v3CDjQuWtiPg9EXvoZmPPgl8APX
-iUgU2vGw8f0i5xE2TxNI7Bkb+gbAjaDf5gpUULLAO6UjCgm8/zcUoMC3OwhmoP0Q
-gh2CrdrppMxJPp96+2KpoZrTropkXPWkAuIFd0QEjuar6FttBcgRgeD3SiP60mNb
-zctEKGqW8ASQU1qYZTU8b8mN4ACdPvNVIwliZyixg4bGCyno/I+VNyvDjME2R56I
-mi+6MnzQycf9tcPPqLn1Lv0ZQDW34b3TBGJKJGDVEnA9iVPDv39zxo/Ft8JjKyWD
-FS5kSvJsUSkgKvsMo5aEMwb80smF0FUZlUoeY7nThfhCqMyvZ/nQHq7EKGY73CwL
-ItPhQdfQHFSXDv2r+Qay7icP8G8Gg3G0zY0/MAC0++aiVboTIEVxFaZ07qh7JwVs
-uOAEnMZv3+qJrs8yXcDDTl7PQw==
------END CERTIFICATE-----
