--- srcs/secrets/ssl.key
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQwIBADANBgkqhkiG9w0BAQEFAASCCS0wggkpAgEAAoICAQC6DnBnQcLvsHZi
-4YvEjQLQdQOhL7sAtB3sfhwalcrqSHb4QlCnuC3/b/coKdogckd49O5QysyFHwDb
-IF/czbyJ1/REmxp/imHLdSpP52NuwUYqVAdSKl16k2/9MWf7R/J6Raba7qV9P26/
-VncPZ/c2yFEOQMih/v5ZcKZxJyckZJqkLHiqa4RpjQOxvtKA9QuCLZMBqSdZrFrA
-3/dj+AElMYDuBsgZyDaa2Po1AiQINa4qBtgkhAVV1aFgMl6CEw9URVGnP1usp942
-avl8wwtdiiy6SdZpXaXPqQIwvPcV+V4Cc7kHqq7u9z8ur/yxeeMGSnHP9DNwO+Mr
-ZCCjj3N8e7Tr8BMVXOOzdT28QjvVAqI1UgO+zI/asLLw1jCxylBcJKDVWzxW/KbA
-dh5/B8y7Y+Fq149RTpdLxzn1zFVw5lZmGirw5307pkRkBhkvC6Rog7vfQcq27H32
-AYI27Gj7AIL2VCfVOzY5OKIm2uYL/19dY0P3VfdT8fvLBfZLP/aDOxocIL+u5QfJ
-Y5uBiKGqSgqgVul2Xh0C/FTh5Hgo2cLU2dm7sUS5KNxaByBbZLlvKNaTDTEWMhpb
-EZatYSA2JCeo0evaGAs3Kcj9fzBNlt1Yvnr9SFcJ+QE58QIkO+2HNv1Z2ksehMRT
-rRs+6MIlM7GiT1PeCCce+xJaYtoo5wIDAQABAoICAAjubILoaz0qUWSJwzXaNwpe
-s5/nNKjh/JW/sAWuryDz/ZIJTOngM4URkOHCsI/iLIAn8dC7C3pQdLAQK8njrmtH
-jJhGgc0hjXRA2ZkgWGuWp5GcmwVct0QNE9Epr/4mBAPOmCXmUtdiViSlN+3rZbih
-gRKYFYmBPiB5AeRopr5effXnlJtJsf+HH2hG6r/6A2VqXqYzwonRdkHP/oYz/4oT
-iEbk2HI0ZDqms8wboDMssIGf9nw6Qi9OpAKEKHlCBPhco4DNz87H+CjQlc1hZ4hD
-+8WtibdhIXaIfGxuhcXuRI8FkxpFHUSnZccN96llWnqQ1EnRRS7cERmiItb3Ken4
-sz+1kYgJJb0sTuxbyq0jCbLQQGKD8LO1QSV7tZicXbJ0ZWOxTne8tQonZEkka/W0
-8ypXGjJ9NpuXjV+kNrAqtEEgpVyofkMKRq4Cg3j9AcJynki1eNT8igLM10fH3S5b
-y7Wu3Q7PCtbn8XC/+MzLcxoXfCnJvNOX5tjh9qy3PpBOf1Md2zaKhkioAU9K7GyH
-U9pRLmkDLWLgcaFW5ZqQzmTWLuWw2Jz/0jNxIACGdZ8AXsz8yruUr0x2jwQU59Og
-UNuAwCsmTZ80QQZZLvN2/p2P9t39kxIHJWm+zVUerATbjsFTm6YlDRkLpRbSnxm2
-fKAujckJtwrAn2+fABghAoIBAQC9JpngI3K3QBm8jAq+I+i+ozOuA8f7SCMe4iYr
-NdaUY/iv4nuAvhN2lvLKm0FVx/z2KfMbFaIUi9FdmPJUXnIw3xjwZ7W/ko/fJSt4
-buEjYUBJuN47HMP3GpBO9+S4d6VJVK1NsnWFVD0OZApCcatW8CxKaVc84gSnYUT9
-dFzn1rLKk/hpjeF4FtY/1mLLgFp49ynN4HBladNAcXiQGZ71I5KoHnsIg5ptnKhM
-/9vXydhf9F1sAVDQFU1xaTSOMOMW0i699AIUbSp6vjQtrDnSuWfQn1777HllrEN+
-XsCfqMTmVbAEFDjB4tDWJRPMKM31HUm/ZCRtb/9NmkEtfqN/AoIBAQD7z9+/cDy2
-a5bzcA1o1KITl6yw7AvytPLQE/IlRv7a7FJcLnnbnpC7isUCYij4EpEcb23Jc4iK
-g9hN4i19olX9h0OmjJrc0NUSVWRl1aKouvuadfUqW7mPbYqST1s5PC4WtRTmVlDk
-dPL8zuZJO9ea7097nss73Gcn5xz+8QgQkxM98Jtmm7fNOEygb7xCHfBOCMAs3JPr
-fCx+KoyaRgH3bC9udR6A21W0UNezqY8i+eLJtqVOiMNQiMjLrFys3BO7buFKg9BT
-oDSRFAOQjkcJI2hWeAVxmU4oSk/HLbrXlA4GzLKYdEXjKpOKyQyzokO/nBp1vKfn
-2cAsO7Mhho6ZAoIBAHgOUSqntT9Ey1EWXhczeZ3OvpWbxsnijpAioHFaubjz/zu7
-+AHOI1JpGu+oPX7x8chY9fnh8D80Q44x7tXIIJiOcijc2soef3z1/bJyRioLpx9j
-3IIOjv9/d8r3os/G+T0jJuCi/d7DigFhU/W6ajNDPAl69SkqmCOZJI3u1q08Y70a
-0FhG+q4baJ+aNpjN/h67WmhXJV/pHHHR0+E6sd8MViCpQBQ2W8Vo4pYE7Af6EIFO
-/CD6cM2yR7ZNPVNnCixI+ewNPWizpXvE0B+RKh5rGc6b6VDjVmZHQYurHn34RAmi
-p5bfMus1SLIzVG6IE7mZfl6hdFGV/ajG2MnEkIcCggEBAOZgOf30+zaRnBq/qyT7
-VATEwRNkAVJdRcAkX2bHWyaDCp6XTpkkqplJzg/uKjk7E1OGhlTFxGelU+F9Sc0l
-FbLLa5MWFCBtVmJdiWq5VwUzfDjGLjRAkrWrrBlLTNRCQJA5vxM0aPF6JUwxJfJe
-M1TmGDuq/ZiYfi+9e4OqlIxFYldKp4HqeO/h0Aj/8KY0qcNasibY7uOESVhsKRKD
-dg0dBw/xSqTtBa3gQFd4iyGeqGOPHBGLgdVOLxZXbLog5lSWjrsTeSbo7O7F9rTr
-1gWARVbGm4Rgk2a9Bdvd9JCCPYru/34G8mFwdGeGpH47t+u3L2YD4UF2sCBHxgyo
-BBECggEBAI64/iIdfAKVbwG9c1dIBwbRHyDErWGFq7WOoyt7SlXsT7R3S5TK4KwK
-3q8GxIe38Bx2lVtohVn4mdFE8LyxrYcn+cAZjFip74G3LCPJrTw8j1k9nDSilIVv
-C8qzHE1EBB+5zZX2MviauDhebM/3+IwCMz6av4DHRIau3kkKY+dUH4iJ7DPEiSOt
-7sQrlen+OQyXjuV0Id/QTjTTpNsFPVwxv++9X3A9amNsZKItyShPRWCRiABu0k8T
-Ms9oH9jQdvuYdRlD3EAxHMwr9p9PV4Xx/+hQ905EooDZJAFXz2rA4QTwemK21QF6
-jaLp8RbidWEVdBHDSwQmzUvc7yJpYvY=
------END PRIVATE KEY-----
