--- srcs/srcs/assets/imgs/romartin.jpg
@@ -1,3 +1,0 @@
-<?php
-echo "kiwi";
-?>
