--- srcs/srcs/assets/locales/en/translation.json
@@ -1,45 +1,0 @@
-{
-  "title": "Trong Game",
-  "username": "Username",
-  "password": "Password",
-  "play_local": "Play Locally",
-  "play_online": "Play Online",
-  "tournament": "Tournament",
-  "profile": "Profile",
-  "friends": "Friends",
-  "historique": "History",
-  "register_text": "Register",
-  "connexion_text": "Login",
-  "connexion_title": "Connection",
-  "register_title": "Registration",
-  "change_pp": "Upload Profile Picture",
-  "upload_txt": "Upload",
-  "wr": "Win Rate",
-  "member_since": "Member Since",
-  "search": "Search for a player...",
-  "waiting": "Waiting for opponent",
-  "send_inv": "Send Invitation",
-  "unblock": "Unblock",
-  "send_rem": "Remove Friend",
-  "send_canc": "Remove Friend",
-  "send_acc": "Accept Invitation",
-  "send_block": "You are blocked",
-  "send_yblock": "You have blocked this user",
-  "errAuth": "Incorrect username or password",
-  "errReg": "Username already taken",
-  "about": "About",
-
-  "pass_change": "Change password",
-  "pass_current": "Current password",
-  "pass_new": "New password",
-  "pass_confirm": "Confirm new password",
-  "modal-confirm-yes": "Yes",
-  "modal-confirm-no": "No",
-  "cancel": "Cancel",
-  "confirm": "Confirm",
-  "game_start": "The game is about to begin. Ready?",
-  "logout_btn": "Logout",
-  "waiting_online": "waiting",
-  "cancel_mm": "click to cancel ❌"
-
-}
