--- srcs/srcs/assets/locales/fr/translation.json
@@ -1,45 +1,0 @@
-{
-  "title": "Jeu Trong",
-  "username": "Nom d'utilisateur",
-  "password": "Mot de passe",
-  "play_local": "Jouer en local",
-  "play_online": "Jouer en ligne",
-  "tournament": "Tournoi",
-  "profile": "Profil",
-  "friends": "Amis",
-  "historique": "Historique",
-  "register_text": "S'inscrire",
-  "connexion_text": "Connexion",
-  "connexion_title": "Connexion",
-  "register_title": "Inscription",
-  "change_pp": "Téléverser une photo de profil",
-  "upload_txt": "Téléverser",
-  "wr": "Taux de victoire",
-  "member_since": "Membre depuis",
-  "search": "Rechercher un joueur...",
-  "waiting": "En attente d’un adversaire",
-  "send_inv": "Envoyer une invitation",
-  "unblock": "Débloquer",
-  "send_rem": "Supprimer l’ami",
-  "send_canc": "Supprimer l’ami",
-  "send_acc": "Accepter l’invitation",
-  "send_block": "Vous êtes bloqué",
-  "send_yblock": "Vous avez bloqué cet utilisateur",
-  "errAuth": "Nom d'utilisateur ou mot de passe incorrect",
-  "errReg": "Ce nom d'utilisateur est déjà utilisé",
-  "about": "À propos",
-
-  "pass_change": "Changer le mot de pass",
-  "pass_current": "Mot de passe actuel",
-  "pass_new": "Nouveau mot de passe",
-  "pass_confirm": "Confirmer le nouveau mot de passe",
-  "modal-confirm-yes": "Oui",
-  "modal-confirm-no": "Non",
-  "cancel": "Annuler",
-  "confirm": "Valider",
-  "game_start": "La partie va commencer. Prêt ?",
-  "logout_btn": "Déconnexion",
-  "waiting_online": "recherche",
-  "cancel_mm": "cliquez pour annuler ❌"
-
-}
