--- srcs/srcs/assets/locales/jp/translation.json
@@ -1,44 +1,0 @@
-{
-  "title": "Trong ゲーム",
-  "username": "ユーザー名",
-  "password": "パスワード",
-  "play_local": "ローカルでプレイする",
-  "play_online": "オンラインでプレイする",
-  "tournament": "トーナメント",
-  "profile": "プロフィール",
-  "friends": "友達",
-  "historique": "履歴",
-  "register_text": "登録する",
-  "connexion_text": "ログイン",
-  "connexion_title": "接続",
-  "register_title": "登録",
-  "change_pp": "プロフィール写真をアップロードする",
-  "upload_txt": "アップロード",
-  "wr": "勝率",
-  "member_since": "参加日",
-  "search": "プレイヤーを検索...",
-  "waiting": "相手を待っています",
-  "send_inv": "招待を送る",
-  "unblock": "ブロック解除",
-  "send_rem": "友達を削除",
-  "send_canc": "友達を削除",
-  "send_acc": "招待を承認",
-  "send_block": "あなたはブロックされています",
-  "send_yblock": "このユーザーをブロックしました",
-  "errAuth": "ユーザー名またはパスワードが正しくありません",
-  "errReg": "このユーザー名はすでに使われています",
-  "about": "このゲームについて",
-
-  "pass_change": "パスワードを変更する",
-  "pass_current": "現在のパスワード",
-  "pass_new": "新しいパスワード",
-  "pass_confirm": "新しいパスワードを確認します",
-  "modal-confirm-yes": "はい",
-  "modal-confirm-no": "いいえ",
-  "cancel": "キャンセル",
-  "confirm": "確認する",
-  "game_start": "ゲームが始まります。準備はいいですか？",
-  "logout_btn": "ログアウト",
-  "waiting_online": "待っている",
-  "cancel_mm": "キャンセルするにはクリックしてください❌"
-}
