--- srcs/srcs/secrets/ssl.crt
@@ -1,32 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFfzCCA2egAwIBAgIUKzxjqitS21a4lBftXiTAM3kKHQIwDQYJKoZIhvcNAQEL
-BQAwTzELMAkGA1UEBhMCRlIxDzANBgNVBAgMBkZyYW5jZTERMA8GA1UEBwwITXVs
-aG91c2UxDTALBgNVBAoMBHBvbmcxDTALBgNVBAMMBG5vbmUwHhcNMjUwNjI1MTI0
-NjE4WhcNMjUwNzI1MTI0NjE4WjBPMQswCQYDVQQGEwJGUjEPMA0GA1UECAwGRnJh
-bmNlMREwDwYDVQQHDAhNdWxob3VzZTENMAsGA1UECgwEcG9uZzENMAsGA1UEAwwE
-bm9uZTCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBALrm+JITdqoeL82Z
-RxaJtOECScZWHy32CMpY52fzht9sGixEPtvnpW/szsWeDsFKWkXJh5TfRwbMuA6T
-Ab3s005Mfm+cyQD4ajUqIWtZns6gFaGl1PuNmCqXYytWh5uZAXf2O19uwqQCKxwx
-aMZYz6xwIBhuprX7IHp+nvDJttxznT3AFLmCSFc2UTfDb4FuQbdTjZnADC9TlXO/
-FF2tpIQSdreNUrMG7NUNE++wVBJfAE5YDOTDdJj38Z/+rPDD77en16UtCVWT6kob
-Zdlv1ML+KOyx2lNRyfkuUz1Aoi1fAvOmq6g5t1UHGVsr6OWzdGB0/mC1Z/K1mRzi
-eCwV26GCgQ9VAE4dv3hg6kMUl4+XWh2wU4KDD0UKuUIuguW6eI1G80RMBo+wNhls
-3ETkBuIE2lt6gMpCsEvLu9cttKj1YOx63vBDD8ulAMVaA37ZbK5vSTZcvUkHVyQj
-Tp8mo7qdqJhbmU74aZJH0vPxSmBWI+SRpLnaxolmRX/FYEqUgckxdH+05sCqj1ZM
-IpGMnMRWwYEycfpdQKVxYTbF9mZhjjZhJQha5JP/6U1dxsHFWi42lX6AaDh6omkd
-xMtymPy3KPe25wetbWVXuqewvlVWf9fqezPQDOfSbmMFDcHN2GDdhCHp4kv/Ma1v
-1GFbcMrzrc1jP/kYUW9JUWpy8deTAgMBAAGjUzBRMB0GA1UdDgQWBBSVUlPj+SRj
-YQTniHShqEEuHJG/mTAfBgNVHSMEGDAWgBSVUlPj+SRjYQTniHShqEEuHJG/mTAP
-BgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4ICAQAvbonx7Yt1sXRY0QPn
-0JhVyQSNsbqSz4SWZxXfd1ZQbRmJ0gXUdRInnxRcbkG9OV0HJV63J/2gBJRbIx4Q
-rIWbCYTzcbJJpTwsaJRM6rJv8ayeSoOb2t1glfEGid2loTCb12bsUTLCRvnib3E9
-Rpi9aPUzCsDev7YQz+UriGVRrWNeITtjibh+/BDWpajaBdJ7Qs6JB72B72VfuBTO
-J7bcv09ZHL/1IpxaVF7wIHHllsLxpvE/NFgjt7vluo5moIByge0PsCgdYrWi2M93
-ZVZ6fF/Wf2pvdoC092PPfGhQdroz35IKIsXevWGetZ6gCZwf0jRNUC6+6N+twXDx
-VunWEZKNifrI0q4B9NmrWR88jVymWgGh2P337m8a+glE45r+ZhV9U1V3adavZhTs
-7IMNU+QVzrPj4izBnOb4HNgczu7MY90dIGwv6Lkl+WO+elA1yh8tirPE/uc1h6IO
-e7x0F8qB+LJE6gQ55O2D02Q02CLRMo832rqBhh0vpn3fEOD9YzsIH4PH0vY64qzB
-McDc1MFQ+S1WgfMJHRl/h3sfMV2MwJ3M5/OTYbM4RhnovkWyiw31/RluyKIfReHM
-awNcl6BUTPnVOFYG9prImLbcXx+ibO+BlepGlciMnDzC8tYVgM7lQqCX30H23TuD
-5N0Jm1M7dOPj3C05CEW4bMYVAw==
------END CERTIFICATE-----
