--- srcs/srcs/secrets/ssl.key
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQwIBADANBgkqhkiG9w0BAQEFAASCCS0wggkpAgEAAoICAQC65viSE3aqHi/N
-mUcWibThAknGVh8t9gjKWOdn84bfbBosRD7b56Vv7M7Fng7BSlpFyYeU30cGzLgO
-kwG97NNOTH5vnMkA+Go1KiFrWZ7OoBWhpdT7jZgql2MrVoebmQF39jtfbsKkAisc
-MWjGWM+scCAYbqa1+yB6fp7wybbcc509wBS5gkhXNlE3w2+BbkG3U42ZwAwvU5Vz
-vxRdraSEEna3jVKzBuzVDRPvsFQSXwBOWAzkw3SY9/Gf/qzww++3p9elLQlVk+pK
-G2XZb9TC/ijssdpTUcn5LlM9QKItXwLzpquoObdVBxlbK+jls3RgdP5gtWfytZkc
-4ngsFduhgoEPVQBOHb94YOpDFJePl1odsFOCgw9FCrlCLoLluniNRvNETAaPsDYZ
-bNxE5AbiBNpbeoDKQrBLy7vXLbSo9WDset7wQw/LpQDFWgN+2Wyub0k2XL1JB1ck
-I06fJqO6naiYW5lO+GmSR9Lz8UpgViPkkaS52saJZkV/xWBKlIHJMXR/tObAqo9W
-TCKRjJzEVsGBMnH6XUClcWE2xfZmYY42YSUIWuST/+lNXcbBxVouNpV+gGg4eqJp
-HcTLcpj8tyj3tucHrW1lV7qnsL5VVn/X6nsz0Azn0m5jBQ3Bzdhg3YQh6eJL/zGt
-b9RhW3DK863NYz/5GFFvSVFqcvHXkwIDAQABAoICAFR2SXG2AwO41ZB7iqEYRRoE
-DzPmuntqOqm6yIR1DIfOEfvcMW3uPpiYcw7iMTfPcoEkJDW2OyJLBIhtjbV3Hzy8
-m8yzlTBvkouJCQY6v2YP+QaOavPtFo100XeFyI2b9Qpxx34Gj0ZPxoyclYVBIESE
-fTSdWJBXtuM0cn5NIep5At+jfJoi0q862DQziSH8Ig2HYC/RlpASnTDr8netH3Ih
-5+1mG6/U2tUi7lX8oNmWfgxphRYicAAq2FId5Kh/W0brPiwicJvwVsn7nKjI/WTL
-N9JTQ18w0G3frfU5zrBjRfG26c8BmiR+K0rEyqB3Wwm2BTjN8fZAc7uWaok9rX0G
-WYXz8IkNeUPHWAkWot1bJszI89D4W5ut6jJ0JBme+bOdvl23LVSvaf3xqbMtJuXS
-s9u2g0F7A/Nb10lIZReGii/lt37t1ePO6TL9Fy55DEh3uDsCEy9KarpD4cOT3eFd
-SFikyWK2/r660GtHQLM3xmhKwGOmPx/5G4GHa0UAO+Y0yM6tzXw52xDi6zHOPgXH
-NJjEMgZK8CsQChUlaSZWiLyjS+RlHrFOcc7LOvRgjjEOXXQ+xLgKiWkEQmw/XV/q
-qKi8NWmmvFIlRVi1bR/ks1XKP38wyBJO+dZV57hH9okmXLo+eV/i+OswfuLaLLPT
-QC12p4BUCnK5IbfljrjNAoIBAQDGcQZFEYmFXCwimQ1aqzERe9DbB25dZO/LqVLv
-DEKXbkcXvBkfObGyp0GK2IduIdjJpkqzlYNDERSqSIYrE33WOhhlWtMqyTy4Hsdl
-gUX3IXI/gD/+WeChn7cycbItK7bsUkFeaRWSucrG7Fz02nEUSj9MQnPRKflv7SPS
-QF5+2J0JPR7k28IplNYCoe4POebJPOZpsLP++k1U9j4/JufGbROhQNFFVP7vPmei
-a7+Jm1PR4g21GLzjXIj2i10ZNgtCHRICXooazfjXh+chg2fQ5VZEJ5w5iGtv58jE
-GsVSg2mEwgJhqyd7oInEtpeN17/1BDKU4bSh9FQjZ3wV6JdlAoIBAQDxHR1/H2lC
-mXU38+DyMwqPmZdDViquGm0FX9f1YG7tb9wnxfgq2cYJSlrp0Zk+gR4v9v9jAnXM
-VrQOs2NEz+gEu/pxHrdH1eXPfUd/prl8IVe3nMk+ygP7J7u3KS117Wd/Tyy5TuEp
-endvabl7bId4kaECHJi11UxuK6FoaQYV0Tlkvp/KaPpwMsU2kUasQkPa2eFtfavs
-yTP/5o3rAL/QPaVbiU1SA9r73pkwvsZe1aVatWXBC9Z6g+3fs5TOsfeL54iaiRr+
-LrfL4AN8fwIYgOLe13wL2vHsWs9qXZfkVZvzuddqNRLaAcaTOn3Pc5JrBp+Zv8zc
-lDQwRzt8US+XAoIBAHRZ7M0ufS2eX5phlvzkhz6zFEB4WPOVrSwCucfDrn/Pfwuh
-lDFEyLXxb3KThsTCP7PsNryZG3+FUlKNd21mXk/vfEJhQJNDrsBd+5e+dmFyUOR+
-tAbUd+sMFIQVOd1WVki16fILN0Tyfryn9dDzC35UEk0KYZHwRHql1mz+pOlG5Lx4
-A9c4Kt3qkJi2jTiMQI8ZkBfELaWA6Tq5rsGkdB1Z8zqHIKzAK0HnmygWeRWb50eW
-fWcDfCzrIrs793hCwdMf7GFLQTp6Lj9hRdnWjFZlQXFwsB7jp+rTQloJQE0jFAlt
-H6xYiCvqcxzN2oqmR7zHm0jEFf8nnPWtRYoq+CkCggEBALNiC+3l+rP348OrB7Jp
-0vOfO3B3jO1W78FN2yfNhPnVfR+CI+4Nc066KQVIURC1kzXwN3T8AZPy9abyQ2qQ
-m3faOqsqRQWG16EU/bq02c7GJ7KNTRQPzAJxPzE3y4CvDY8j4oxDaTUoze7v43U1
-+1vyTK17hQnTFTfp8f3yFPrqXB7dy5gVni0XOK4uSzwhWMZbz2FcXCqgTrWWXAN4
-3TwagQmnDwVzpNo6dBJV/NPVc14XEwsauYkQ2krSGmXk9dz6dyEx3Ak7zLY7hZx8
-CZ0dVqyf+MQBMne74BCsbTMNlicwHln//nR/WbNzkuJig45osRj8H3gEMwewF0uY
-QyUCggEBAItp3g1vLJMjM/j3CUZvBU58tpkttDWAR9zpDPVkwqRXdnlWMNhESnOD
-RZMMpguVsYK7knFGn8w/aEqUK05Y+alQQUrLQYeLz6UcXOZ9uLh4iPdkr0aE+wKL
-Fx3dnfajOPyiKEaLGbQ9PSPq91yUdCW8TRiYZoghDLQr6FYAoJTiDufntZy5DC8p
-TgY6qsQJgUlb35ebddqny9dMyg3vabl/p78D3wikNakv178xHkILWYCQO1BBJ3th
-fa2YMomdeATjjk1aweFN+CwQyczBK9+NGw+ofFVWZ7A2td1VEuP9NAK+BGxWYWCo
-1JRRgaxjvdB2gXKDZ99BoDx0UUvYyBI=
------END PRIVATE KEY-----
