--- srcs/srcs/services/fastify/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "transcendance",
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/server.js",
-  "scripts": {
-    "buildcss": "npx @tailwindcss/cli -i ./src/input.css -o ./dist/style.css --watch ",
-    "compile": "npx tsc",
-    "start": "node --watch dist/server.js",
-    "run": "node --watch dist/server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "module",
-  "dependencies": {
-    "@fastify/cookie": "^11.0.2",
-    "@fastify/jwt": "^9.1.0",
-    "@fastify/multipart": "^9.0.3",
-    "@fastify/static": "^8.1.1",
-    "@fastify/view": "^10.0.2",
-    "@fastify/websocket": "^11.0.2",
-    "@tailwindcss/cli": "^4.1.4",
-    "bcrypt": "^5.1.1",
-    "better-sqlite3": "^11.10.0",
-    "ejs": "^3.1.10",
-    "fastify": "^5.3.2", 
-    "fastify-bcrypt": "^1.0.1",
-    "fastify-cookie": "^5.6.1",
-    "fastify-multipart": "^5.3.1",
-    "fastify-plugin": "^5.0.1",
-    "jsonwebtoken": "^9.0.2",
-    "socket.io": "^4.8.1",
-    "tailwindcss": "^4.1.4",
-    "fastify-metrics": "^10.x"
-  },
-  "devDependencies": {
-    "@types/node": "^22.13.4",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.7.3"
-  }
-}
