--- srcs/srcs/services/fastify/src/TournamentController.ts
@@ -1,395 +1,0 @@
-import { TournamentClientSocket } from "./TournamentClientSocket.js";
-import { SiteController } from "./SiteController.js";
-
-export class Tournament {
-    private id : number = -1;
-    private name : string = "placeholder";
-    private owner : string = "unowned";
-    private players; // image, win rate{}
-    private is_started : boolean = false;
-
-
-    constructor(tournament) {
-        this.id = tournament.id;
-        this.name = tournament.name;
-        this.owner = tournament.owner;
-        this.players = tournament.players;
-    }
-
-    getId() {
-        return (this.id);
-    }
-
-    getName() {
-        return (this.name);
-    }
-
-    getOwner() {
-        return (this.owner);
-    }
-
-    getPlayers(){
-        return (this.players);
-    }
-
-    isStarted(){
-        return (this.is_started);
-    }
-
-    update(tournament){
-        this.players = tournament.players;
-    }
-}
-
-export class TournamentController {
-
-    /**CONTROLLER */
-    private cws: TournamentClientSocket = null;
-    private site: SiteController = null;
-    private username: string = null;
-    private tournament: Tournament = null;
-
-    /**VIEW */
-    private tournament_page = document.getElementById("tournament_page");
-    private tournaments_page = document.getElementById("tournaments_join");
-    private tournament_div = document.getElementById("tournament");
-    private tournament_state = document.getElementById("tournament_state");
-    private tournaments_list = document.getElementById("tournaments_list");
-    private tournament_form = document.getElementById("tournament_form");
-    private tournament_create_btn = document.getElementById("tournament_create_button");
-    private tournament_join_btn = document.getElementById("tournament_join_button");
-    private tournament_rejoin_btn = document.getElementById("tournament_rejoin_button");
-
-// need to tell the client if a tournament created by him already exists
-// maybe the client send a fetch to check before creating the socket ?
-// Need to check (on page load) if the client has a tournament already present (maybe with the connection socket, UserSocket (profileSocket))
-
-// Need to handle if the tournament is disbanded
-    constructor(site) {
-        this.site = site;
-        // fetch to check if a tournament already exists
-        // this.checkTournament();
-    }
-
-    setUsername(username) {
-        this.username = username;
-    }
-
-    hasTournament(){
-        return (this.tournament !== null);
-    }
-
-    addEvents() {
-        this.tournament_create_btn.addEventListener("click", (event) => {
-            event.preventDefault();
-
-            // if (this.tournament !== null && this.tournament !== undefined){
-            if (this.tournament !== null) {
-                // print_error("You're already registered for a tournament");
-                alert("You're already registered for a tournament");
-                return;
-            }
-            this.site.hide_all();
-
-            this.print_tournament_page();
-            this.print_tournament_form();
-        })
-
-        this.tournament_form.addEventListener("submit", async (event) => {
-            event.preventDefault();
-            const name = document.getElementById("tournament_name") as HTMLInputElement;
-            //Verifier que l'input est valide avant de l'envoyer !
-            if (name.value === null || name.value === ""){
-                alert("cant be empty");
-                return ;
-            }
-            try {
-                const body = {
-                    owner: this.username,
-                    tournament_name: name.value
-                }
-                const resp = await fetch('/tournament/create', {
-                    method: 'POST',
-                    headers: { "Content-Type": "application/json" },
-                    body: JSON.stringify(body)
-                });
-                const data = await resp.json();
-                if (data.success) { 
-                    this.tournament = new Tournament(data.tournament);
-                    this.cws = new TournamentClientSocket(this.username, this, this.tournament);
-                    // console.log(this.tournament);
-                    this.hide_all();
-                    this.print_tournament();
-                    this.print_tournament_page();
-                    this.print_tournament_rejoin_btn();
-                }
-                else
-                    throw (Error(data.error));
-            } catch (error) {
-                console.log("error: ", error);
-            }
-        });
-
-        this.tournament_join_btn.addEventListener("click", async (event) => {
-            event.preventDefault();
-
-            this.site.hide_all();
-
-            this.print_tournament_page();
-            this.print_tournaments_page();
-            this.clear_tournaments();
-            try {
-                const resp = await fetch('/tournament/list?username=' + this.username, {
-                    method: 'GET',
-                    headers: { "Content-Type": "application/json" }
-                });
-                const data = await resp.json();
-
-                // console.log(data);
-
-                if (data.success) {
-                    if (data.tournaments !== null && data.tournaments !== undefined) {
-                        let i = 0, size = data.tournaments.length;
-                        let list = document.createElement("ul");
-                        if (size == 0)
-                            this.tournaments_list.append(document.createTextNode("No tournament found. Try creating one !"));
-                        else {
-                            list.appendChild(document.createTextNode("List of tournaments available"));
-                            while (i < size) {
-                                let el = document.createElement("li");
-                                el.appendChild(document.createTextNode(data.tournaments[i].name));
-                                el.appendChild(document.createTextNode(data.tournaments[i].players.length + "/" + "8"));
-                                el.setAttribute("tournament_id", data.tournaments[i].id);
-                                el.addEventListener("click", async (event) => {
-                                    event.preventDefault();
-                                    try {
-                                        let query = new URLSearchParams();
-                                        query.append("username", this.username);
-                                        let url = '/tournament/join/' + (event.target as Element).getAttribute("tournament_id") + `?${query}`;
-                                        const resp = await fetch(url, {
-                                            method: 'GET',
-                                            headers: { "Content-Type": "application/json" }
-                                        });
-                                        const data = await resp.json();
-                                        if (data.success) {
-                                            this.tournament = new Tournament(data.tournament);
-                                            this.cws = new TournamentClientSocket(this.username, this, this.tournament);
-
-                                            // console.log(this.tournament);
-
-                                            this.site.hide_all();
-                                            this.clear_tournament();
-                                            this.print_tournament_page();
-                                            this.print_tournament();
-                                        } else
-                                            throw Error(data.error);
-                                    } catch (error) {
-                                        console.log(error);
-                                    }
-                                });
-                                list.append(el);
-                                i++;
-                            }
-                            this.tournaments_list.appendChild(list);
-                        }
-                    } else
-                        throw (Error(data.error));
-                }
-            } catch (err) {
-                console.log(err);
-            }
-        });
-
-        this.tournament_rejoin_btn.addEventListener("click", async (event) => {
-            this.clear_tournament();
-            this.site.hide_all();
-            this.print_tournament_page();
-            this.print_tournament();
-        });
-    }
-
-    updateTournament(tournament : Tournament){
-        this.tournament.update(tournament);
-        this.clear_tournament();
-        this.site.hide_all();
-        this.print_tournament();
-        this.print_tournament_page();
-    }
-
-    finishGame(){
-        
-    }
-    /**
-     * VIEW METHODS
-     * 
-     */
-    print_menu(){
-        this.site.print_menu();
-    }
-
-    print_tournament_page() {
-        this.tournament_page.classList.replace("hidden", "flex");
-    }
-
-    hide_tournament_page() {
-        this.tournament_page.classList.replace("flex", "hidden");
-    }
-
-    print_tournaments_page() {
-        this.tournaments_page.classList.replace("hidden", "flex");
-    }
-
-    hide_tournaments_page() {
-        this.tournaments_page.classList.replace("flex", "hidden");
-    }
-
-    print_on_going_tournament(){
-        this.tournament_state.classList.replace("hidden", "flex");
-    }
-
-    hide_on_going_tournament(){
-        this.tournament_state.classList.replace("flex", "hidden");
-    }
-
-    print_tournament() {
-        console.log("Printing tournament");
-        console.log(this.tournament);
-        if (this.tournament === null){
-            alert("Not implemented yet (print tournament but tournament is null)");
-            return ;
-        } else if (this.tournament.isStarted()){
-            this.print_on_going_tournament();
-            return ;
-        }
-        this.tournament_div.classList.replace("hidden", "block");
-
-        let title = document.createElement("h3");
-        title.append(document.createTextNode(this.tournament.getName()));
-
-        let table = document.createElement("table");
-        table.append();
-
-        //First line
-        let tr = document.createElement("tr");
-        let th = document.createElement("th");
-
-        th.append(document.createTextNode("User"));
-        tr.append(th);
-
-        th = document.createElement("th");
-        th.append(document.createTextNode("Role"));
-        tr.append(th);
-
-        table.append(tr);
-
-        //Next lines
-        this.tournament.getPlayers().forEach(p => {
-            tr = document.createElement("tr");
-            th = document.createElement("th");
-
-            th.append(document.createTextNode(p));
-            tr.append(th);
-
-            th = document.createElement("th");
-            if (p === this.tournament.getOwner()) {
-                th.append(document.createTextNode("Owner"));
-            }
-            else {
-                th.append(document.createTextNode("Player"));
-            }
-            tr.append(th);
-            table.append(tr);
-        });
-
-        this.tournament_div.append(title);
-        this.tournament_div.append(table);
-
-        if (this.username === this.tournament.getOwner()) {
-            let start_button = document.createElement("button");
-            start_button.append(document.createTextNode("Start"));
-            start_button.onclick = (event) => this.startTournamentHandler(event);
-            this.tournament_div.append(start_button);
-        }
-
-        let leave_button = document.createElement("button");
-        leave_button.append(document.createTextNode("Leave"));
-
-        leave_button.onclick = (event) => this.leaveTournamentHandler(event);
-        this.tournament_div.append(leave_button);
-    }
-
-    async leaveTournamentHandler(event) {
-        event.preventDefault();
-        // console.log("Trying to leave so soon ?");
-
-        try {
-            // fetch HERE TODO
-            let query = new URLSearchParams();
-            query.append("username", this.username);
-            console.log("Trying to leave ");
-            console.log(this.tournament);
-            let url = '/tournament/leave/' + this.tournament.getId() + `?${query}`;
-
-            const resp = await fetch(url, {
-                method: 'GET'
-            });
-
-            const data = await resp.json();
-            if (data.success) {
-                this.tournament = null;
-                this.hide_all();
-                this.cws.close();
-                this.cws = null;
-                this.print_tournament_page();
-            } else {
-                console.log("Didnt leave");
-                throw (Error(data.error));
-            }
-        } catch (error) {
-            console.log(error);
-        }
-    }
-
-    async startTournamentHandler(event) {
-        this.cws.startTournament();
-    }
-
-    hide_tournament() {
-        this.tournament_div.classList.replace("block", "hidden");
-    }
-
-    print_tournament_form() {
-        this.tournament_form.classList.replace("hidden", "flex");
-    }
-
-    hide_tournament_form() {
-        this.tournament_form.classList.replace("flex", "hidden");
-    }
-
-    print_tournament_rejoin_btn(){
-        this.tournament_rejoin_btn.classList.replace("hidden", "flex");
-    }
-
-    hide_tournament_rejoin_btn(){
-        this.tournament_rejoin_btn.classList.replace("flex", "hidden");
-    }
-
-    clear_tournament() {
-        // console.log(this.tournament_div.textContent);/
-        this.tournament_div.textContent = '';
-    }
-
-    clear_tournaments() {
-        // console.log(this.tournaments_list.textContent);
-        this.tournaments_list.textContent = '';
-    }
-
-    hide_all() {
-        this.hide_tournament_page();
-        this.hide_tournaments_page();
-        this.hide_tournament_form();
-        this.hide_tournament_rejoin_btn();
-        this.hide_tournament();
-        this.hide_on_going_tournament();
-    }
-};
