--- srcs/srcs/services/fastify/src/input.css
@@ -1,1 +1,0 @@
-@import "tailwindcss";
