--- srcs/srcs/services/fastify/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "dist",
-    "module": "nodenext"
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules"]
-}
